Move all dashboard-related components to page-specific components folder
