+++ frontend/src/pages/protected/Dashboard/components/DashboardNavbar.tsx
