+++ frontend/src/pages/protected/Dashboard/components/DashboardSuggestions.tsx
